Introduce a ResizableBox component and apply to image/spacer blocks
