+++ packages/block-library/src/image/edit.js
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import classnames from 'classnames';
-import ResizableBox from 're-resizable';
 import {
 	get,
 	isEmpty,
@@ -22,6 +21,7 @@ import {
 	ButtonGroup,
 	IconButton,
 	PanelBody,
+	ResizableBox,
 	SelectControl,
 	TextControl,
 	TextareaControl,
@@ -446,21 +446,10 @@ class ImageEdit extends Component {
 							}
 							/* eslint-enable no-lonely-if */
 
-							// Removes the inline styles in the drag handles.
-							const handleStylesOverrides = {
-								width: null,
-								height: null,
-								top: null,
-								right: null,
-								bottom: null,
-								left: null,
-							};
-
 							return (
 								<Fragment>
 									{ getInspectorControls( imageWidth, imageHeight ) }
 									<ResizableBox
-										className="block-library-image__resizer"
 										size={
 											width && height ? {
 												width,
@@ -472,16 +461,6 @@ class ImageEdit extends Component {
 										minHeight={ minHeight }
 										maxHeight={ maxWidth / ratio }
 										lockAspectRatio
-										handleClasses={ {
-											right: 'block-library-image__resize-handler-right',
-											bottom: 'block-library-image__resize-handler-bottom',
-											left: 'block-library-image__resize-handler-left',
-										} }
-										handleStyles={ {
-											right: handleStylesOverrides,
-											bottom: handleStylesOverrides,
-											left: handleStylesOverrides,
-										} }
 										enable={ {
 											top: false,
 											right: showRightHandle,
