+++ packages/block-library/src/spacer/editor.scss
@@ -1,22 +1,3 @@
 .block-library-spacer__resize-container.is-selected {
-	.block-library-spacer__resize-handler-bottom {
-		display: block;
-	}
-}
-
-.block-library-spacer__resize-container.is-selected {
 	background: $light-gray-200;
-}
-
-.block-library-spacer__resize-handler-bottom {
-	@include resize-handler__container();
-
-	// Offset the handle container's position.
-	position: absolute;
-	bottom: calc(#{$resize-handler-container-size / 2} * -1);
-	left: calc(50% - #{$resize-handler-container-size / 2});
-}
-
-.block-library-spacer__resize-handler-bottom::before {
-	@include resize-handler__visible-handle();
 }
