+++ packages/components/package.json
@@ -40,6 +40,7 @@
 		"memize": "^1.0.5",
 		"moment": "^2.22.1",
 		"mousetrap": "^1.6.2",
+		"re-resizable": "^4.7.1",
 		"react-click-outside": "^2.3.1",
 		"react-color": "^2.13.4",
 		"react-datepicker": "^1.4.1",
