+++ packages/components/src/resizable-box/README.md
@@ -1,0 +1,5 @@
+# ResizableBox
+
+ResizableBox is a wrapper around the [re-resizable package](https://github.com/bokuweb/re-resizable) with pre-defined classes and styles.
+
+
