+++ packages/components/src/style.scss
@@ -27,6 +27,7 @@
 @import "./popover/style.scss";
 @import "./radio-control/style.scss";
 @import "./range-control/style.scss";
+@import "./resizable-box/style.scss";
 @import "./responsive-wrapper/style.scss";
 @import "./sandbox/style.scss";
 @import "./scroll-lock/style.scss";
